New file: FindHotPixels.py
Added new file which should be used to find the hot pixels of a run.
It's simpler to make the hot pixel file once (here) and then only load
it in ComputeAlignment, pyEudetReconstructionOnly etc.

The file takes regular inputs, and simply calls FindHotPixel.
Control plots are saved. The usual usage would be to call this file,
then ComputeAlignment, then the reconstruction and analysis.