Possibly improved TotalRotationFunction and Perform3StepAlignment
The reason 'possibly' is because while I believe the method is improved,
the results are actually very similar to before. But now atleast there is
the possibility to find rotations around X and Y. Even though these seem
small in practice.

In Perform3StepAlignment:
- removed unnecessary gtol parameter from options
- altered BFGS to better behaving Nelder-Mead for minimisation of TotalRotationFunction
This doesn't need a step size as input, so it's more versatile. In testing with
6 different runs it behaves well.

In TotalRotationFunction:
- removed unnecessary variables
- improved plot ranges
- added a fitted sigma - seems to make things more precise / less unstable

In ComputeAlignment:
- removed gtol from the call to Perform3StepAlignment