Simplification of EtaCorrection code to only use one sigma.
Since we have no B field or angle tracks, the EtaCorrection sigma
should always be equal in x and y. So only use one sigma parameter.

This actually makes no difference to current results, since sigmaX
and sigmaY are currently forced to be the same.

Shir's note (LCD-OPEN-2014-001) shows that when these two parameters
are allowed to be different, they consistently come out to be the
same, as expected.