Better TotalDistanceFunction for the alignment procedure.
Reversed the emphasis, so that instead of aligning clusters to the
nearest track, tracks are aligned to the nearest cluster. This is
more correct, as there are more clusters than tracks due to the high
hit efficiency. Have checked the effect on the Timepix data and
results are consistent / better.