Analysing edge efficiency independent of edge width.
Since the efficiency across the edge of an assembly is interesting for
all assemblies, remove the check that the edge width is non-zero and make
efficiency plots as a function of distance from the edge.

In ToolBox.py:
- change limits of histograms to go over the edge
- remove IsInEdges requirement
- look for tracks close to the edge instead

In pyEudetAnalysisOnly.py:
- remove the non-zero edge requirement

Note that for this code to be sucessful, it must currently run with
parameter -e 0.07 (for instance). Using -e 0.0 will prevent matching of
tracks in the edges (I think) and so the edge efficiency does to zero
immediately. Will try to work on this.