Since we define halfChip_X and Y, use them.
Replaced npix_X*pitchX/2. with halfChip_X in all files. This alteration
has no effect on results, but makes code easier to read.

Also changed function IsInGoodRegion (which was never called and was doing
something highly specific) into IsInMain, which will complement IsInEdges.
This new function is yet to be used but will be necessary for future edits
to the efficiency calculations.