Tracks can be outside the DUT and still be matched.
In FindMatchedCluster, remove the requirement that the track is inside the
DUT (or the DUT+edges) to be matched. Efficiency may (does) extend beyond the
edges, and should be tested for.

This means the efficiencies have to be calculated abit more carefully. This is
done, checking tracks are in main (IsInMain) or edges (IsInEdges) before their
matched cluster contributing to the efficiency.